fix(service-worker): install service worker in Init component
