+++ src/Bootstrap.tsx
@@ -7,7 +7,6 @@ import {
 } from 'react-router-dom'
 import localforage from 'localforage'
 
-import * as serviceWorkerRegistration from 'serviceWorkerRegistration'
 import { StorageContext } from 'contexts/StorageContext'
 import { SettingsContext } from 'contexts/SettingsContext'
 import { homepageUrl, routes } from 'config/routes'
@@ -33,6 +32,7 @@ import {
 } from 'services/Serialization'
 
 export interface BootstrapProps {
+  appNeedsUpdate: boolean
   persistedStorage?: typeof localforage
   initialUserSettings: UserSettings
   serializationService?: typeof serializationServiceInstance
@@ -82,6 +82,7 @@ const Bootstrap = ({
     description: 'Persisted settings data for chitchatter',
   }),
   initialUserSettings,
+  appNeedsUpdate,
   serializationService = serializationServiceInstance,
 }: BootstrapProps) => {
   const queryParams = useMemo(
@@ -90,15 +91,10 @@ const Bootstrap = ({
   )
 
   const [persistedStorage] = useState(persistedStorageProp)
-  const [appNeedsUpdate, setAppNeedsUpdate] = useState(false)
   const [hasLoadedSettings, setHasLoadedSettings] = useState(false)
   const [userSettings, setUserSettings] =
     useState<UserSettings>(initialUserSettings)
   const { userId } = userSettings
-
-  const handleServiceWorkerUpdate = () => {
-    setAppNeedsUpdate(true)
-  }
 
   const persistUserSettings = useCallback(
     async (newUserSettings: UserSettings) => {
@@ -116,10 +112,6 @@ const Bootstrap = ({
     },
     [persistedStorageProp, queryParams, serializationService, userSettings]
   )
-
-  useEffect(() => {
-    serviceWorkerRegistration.register({ onUpdate: handleServiceWorkerUpdate })
-  }, [])
 
   useEffect(() => {
     ;(async () => {
